The first test to validate an entry (#293)
