+++ git-index/Cargo.toml
@@ -23,6 +23,7 @@ git-hash = { version ="^0.8.0", path = "../git-hash" }
 quick-error = "2.0.0"
 memmap2 = "0.5.0"
 filetime = "0.2.15"
+bstr = { version = "0.2.13", default-features = false, features = ["std"]}
 
 serde = { version = "1.0.114", optional = true, default-features = false, features = ["derive"] }
 smallvec = "1.7.0"
